+++ InfinityHammer/config/ConfigurationBlueprint.cs
@@ -1,12 +1,8 @@
 using System.Collections.Generic;
-using System.Globalization;
 using System.IO;
-using System.Linq;
 using BepInEx;
 using BepInEx.Configuration;
-using ServerDevcommands;
 using Service;
-using UnityEngine;
 namespace InfinityHammer;
 
 public partial class Configuration